# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Film-collection API",
    "contact": {
      "name": "Yuliia Kopysova",
      "email": "julykopysova97@b.com"
    }
  },
  "host": "172.20.0.3:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
      "/users": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                  "402":{
                      "description": "You aren't admin"
                  }
                },
                "summary": "Get users",
                "description": ":return: Response",
                "tags": [
                    "Users"
                ]
            },
            "post": {
                "summary": "API Endpoint for create user by admin",
                "description": "Create new user by admin",
                "tags": [
                            "Users"
                        ],
                "parameters": [
                  {
                    "name": "user",
                    "description": "User params for creating",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "New user has created"
                  },
                  "401": {
                    "description": "You aren't admin"
                  }
                }
            }
        },
    "/users/{user_id}": {
            "parameters": [
                {
                    "name": "user_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "204": {
                        "description": "User was deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "You aren't admin"
                    }
                },
                "summary": "Delete user",
                "description": ":param user_id: id of user\n:return: Response",
                "tags": [
                    "Users"
                ]
            },
            "patch": {
              "parameters": [
                  {
                    "name": "user",
                    "description": "User params for updating",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "You aren't admin or this user"
                    }
                },
                "summary": "Updates a user",
                "description": ":param user_id: id of user\n:return: JSON",
                "tags": [
                    "Users"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                  "401": {
                        "description": "You aren't admin or this user"
                    }
                },
                "summary": "Get one user",
                "description": ":param user_id: id of user\n:return: JSON",
                "tags": [
                    "Users"
                ]
            }
        },
...
      "/genres_list": {
        "get": {
          "parameters": [
                  {
                    "name": "start",
                    "description": "from what items start validation",
                    "in": "query",
                    "schema" : {

                        }
                  },
                  {
                    "name": "limit",
                    "description": "how many items in page",
                    "in": "query",
                    "schema" : {

                        }
                  }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          },
          "summary": "Get genres",
          "description": ":return: JSON",
          "tags": [
            "Genres"
          ]
        }
      },
      "/search_film": {
        "get": {
          "parameters": [
                  {
                    "name": "name",
                    "description": "search by part title",
                    "in": "query",
                    "schema" : {

                        }
                  }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          },
          "summary": "Search film by partial coincidence",
          "description": ":return: JSON",
          "tags": [
            "Films"
          ]
        }
      },
    "/films_years_filter": {
        "get": {
          "parameters": [
                  {
                    "name": "year_start",
                    "description": "lower limit of value lower",
                    "in": "query",
                    "schema" : {

                        }
                  },
                  {
                    "name": "year_end",
                    "description": "upper limit of value",
                    "in": "query",
                    "schema" : {

                        }
                  }

          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          },
          "summary": "Filters films between years",
          "description": ":return: JSON",
          "tags": [
            "Films"
          ]
        }
      },
  "/films_director_filter": {
        "get": {
          "parameters": [
                  {
                    "name": "first_name",
                    "description": "first name of director",
                    "in": "query",
                    "schema" : {

                        }
                  },
                  {
                    "name": "last_name",
                    "description": "last name of director",
                    "in": "query",
                    "schema" : {

                        }
                  }

          ],
          "responses": {
            "200": {
              "description": "Success, No suitable films exist"
            },
            "401": {
              "description": "Enter first name and last name"
            }
          },
          "summary": "Filters films by director",
          "description": ":return: JSON",
          "tags": [
            "Films"
          ]
        }
      },
  "/films_genre_filter": {
        "get": {
          "parameters": [
                  {
                    "name": "genre_title",
                    "description": "first name of director",
                    "in": "query",
                    "schema" : {

                        }
                  }

          ],
          "responses": {
            "200": {
              "description": "Success"
            },
            "401": {
              "description": "Enter genre title"
            }
          },
          "summary": "Filters films by genre",
          "description": ":return: JSON",
          "tags": [
            "Films"
          ]
        }
      }
  }
}